Auto-create object hierarchies and arrays
This makes it far easier to create complex hierarchies by just setting a leaf property in an entire object graph, whether by navigating via properties or indexers.

Array items can be added at the end of an existing array by specifying the desired index, or either `-1` or `^1` (for the C# range index notation meaning "first from last") to add to the end.

Fixes #16